+++ src/providers/nvmPosix.js
@@ -52,6 +52,10 @@ module.exports = {
     return `nvm install ${version}`;
   },
 
+  uninstallCommand(version) {
+    return `nvm uninstall ${version}`;
+  },
+
   binDir(version) {
     return path.join(versionsDir(), `v${version}`, 'bin');
   }